src/main/java/io/github/deogicorgi/spring/jpa/service/MemberService.java
package io.github.deogicorgi.spring.jpa.service;

import io.github.deogicorgi.spring.jpa.entity.Member;
import io.github.deogicorgi.spring.jpa.repository.MemberRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
@RequiredArgsConstructor
public class MemberService {
    private final MemberRepository memberRepository;

    public Member save(Member member) {
        return memberRepository.save(member);
    }

    public Member findById(Integer id) {
        Optional<Member> member = memberRepository.findById(id);
        return member.orElseThrow();
    }
}
